Agrega proteccion CSRF a las pantallas de confirmar eliminacion
Los links de "Eliminar" en 9 pantallas (roles, profesionales, pacientes,
cardiologia, empleados, afiliados, historia clinica, lista de espera,
movimientos de caja) no validaban un token CSRF antes de borrar, solo
el parametro confirmar=si. De paso se corrige historia_clinica_delete.php,
que usaba mysql_query (funcion inexistente en PHP moderno) con SQL
armado por concatenacion.

+++ sistema/secciones/afiliados_delete.php
@@ -16,12 +16,15 @@
         <div class="panel-body">
         <?php
 
+          require_once __DIR__ . '/../inc/csrf.php';
+
           $confirmar = $_GET['confirmar'] ?? '';
           $id  = $_GET['id']  ?? 0;
           $pid = $_GET['pid'] ?? 0;
           $rand = $_GET['nc'] ?? '';
+          $tokenValido = hash_equals($_SESSION['csrf_token'] ?? '', $_GET['csrf_token'] ?? '');
 
-          if($confirmar == 'si'){
+          if($confirmar == 'si' && $tokenValido){
 
             $stmt = $pdo->prepare("
                 DELETE FROM pagos_afiliados
@@ -46,7 +49,7 @@
               Esta acci&oacute;n no puede deshacerse.<br>
             </div>
             <div class="pull-right">
-              <a href="?seccion=afiliados_delete&pid='.$pid.'&id='.$id.'&confirmar=si&nc='.$rand.'" class="btn btn-info">Eliminar</a>
+              <a href="?seccion=afiliados_delete&pid='.$pid.'&id='.$id.'&confirmar=si&csrf_token='.urlencode(csrf_token()).'&nc='.$rand.'" class="btn btn-info">Eliminar</a>
               <a href="?seccion=socios_historial&id='.$id.'&nc='.$rand.'" class="btn btn-info">Cancelar</a>
             </div>';
           }
